♻️ JE-103-refactor: map데이터 생성 위치 페이지 컴포넌트로 변경
